+++ tests/Functional/ArrayDtoTest.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the API Platform project.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace ApiPlatform\Tests\Functional;
+
+use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
+
+final class ArrayDtoTest extends ApiTestCase
+{
+    public function testWithGroupFilter(): void
+    {
+        $response = self::createClient()->request('GET', '/array_property_dto_operations');
+        $this->assertArraySubset(['name' => 'test', 'greetings' => [['name' => 'test2']]], $response->toArray());
+    }
+}